# src/main/resources/openapi/openapi.yaml
{
   "openapi":"3.0.1",
   "info":{
      "title":"OpenAPI definition",
      "version":"v0"
   },
   "servers":[
      {
         "url":"http://localhost:8080",
         "description":"Generated server url"
      }
   ],
   "paths":{
      "/api/v1/commercial-activities":{
         "post":{
            "tags":[
               "Commercial Activity API V1"
            ],
            "summary":"Create a new commercial activity",
            "operationId":"createCommercialActivity",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/CommercialActivityRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK"
               }
            }
         }
      },
      "/api/v1/commercial-activities/search":{
         "post":{
            "tags":[
               "Commercial Activity API V1"
            ],
            "summary":"Search similar commercial activities by town and categories in",
            "operationId":"searchSimilarCommercialActivitiesByTown",
            "parameters":[
               {
                  "name":"numberOfResults",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"integer",
                     "format":"int32",
                     "default":10
                  }
               },
               {
                  "name":"town",
                  "in":"query",
                  "required":true,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "requestBody":{
               "content":{
                  "multipart/form-data":{
                     "schema":{
                        "required":[
                           "text"
                        ],
                        "type":"object",
                        "properties":{
                           "text":{
                              "type":"string"
                           },
                           "media":{
                              "type":"array",
                              "items":{
                                 "type":"string",
                                 "format":"binary"
                              }
                           }
                        }
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/CommercialActivityResponse"
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/v1/commercial-activities/batch":{
         "post":{
            "tags":[
               "Commercial Activity API V1"
            ],
            "summary":"Create multiple commercial activity",
            "operationId":"createCommercialActivities",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "type":"array",
                        "items":{
                           "$ref":"#/components/schemas/CommercialActivityRequest"
                        }
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK"
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "CommercialActivityRequest":{
            "type":"object",
            "properties":{
               "taxCode":{
                  "type":"string"
               },
               "name":{
                  "type":"string"
               },
               "town":{
                  "type":"string"
               },
               "address":{
                  "type":"string"
               },
               "categories":{
                  "type":"array",
                  "items":{
                     "type":"string"
                  }
               },
               "description":{
                  "type":"string"
               }
            }
         },
         "CommercialActivityResponse":{
            "type":"object",
            "properties":{
               "taxCode":{
                  "type":"string"
               },
               "name":{
                  "type":"string"
               },
               "town":{
                  "type":"string"
               },
               "address":{
                  "type":"string"
               },
               "categories":{
                  "type":"array",
                  "items":{
                     "type":"string"
                  }
               },
               "description":{
                  "type":"string"
               },
               "score":{
                  "type":"number",
                  "format":"double"
               }
            }
         }
      }
   }
}